Renamed Search configuration file
