+++ src/file_processor.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import Dict, List
 
 from src.matching import SimpleMatching
-from src.named_patterns import SearchConfiguration
+from src.search_configuration import SearchConfiguration
 from src.text_file import TextFile
 from src.text_processor import TextProcessor
